+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Next]
 
+### Changed
+
+- Absinthe plugs use a map instead of custom struct as context
+
 ## [0.4.2] - 2022-06-01
 
 ### Fixed